# backend/docs/mobile/openapi.yaml
openapi: 3.1.0
info:
  title: HR Workforce — Mobile API
  version: 1.0.0
  description: |
    Contract ของ Native App (Expo Employee App)

    Endpoint ทุกตัวเป็น façade อยู่ภายใน backend เดิม ไม่ใช่ backend แยก (ADR-001)
    ดูคู่ mapping ไป Service เดิมได้ที่ `docs/mobile/API_MAPPING_MATRIX.md`

    **เวอร์ชัน:** ขึ้น `/api/mobile/v2` เมื่อมี breaking change เท่านั้น
    การเพิ่ม field ใหม่ถือว่า backward-compatible และทำได้ใน v1
servers:
  - url: /api/mobile/v1
...
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: เข้าสู่ระบบจากแอป
      description: |
        **2FA บนช่องทางมือถือถูกปิดไว้เป็นค่าเริ่มต้น (BE-MOB-002)**

        ระบบยังไม่มีช่องทางส่ง OTP ถ้าบังคับ ผู้ใช้ที่เข้าเงื่อนไขจะล็อกอินไม่ได้ถาวร
        `requiresTwoFactor` จึงเป็น `false` เสมอจนกว่าจะตั้ง `MOBILE_TWO_FACTOR_ENABLED=true`

        แอปต้องรองรับทั้งสองกรณีอยู่แล้ว เพราะ flag เปิดคืนได้โดยไม่ต้องออกเวอร์ชันใหม่
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, installation]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
                installation: { $ref: '#/components/schemas/Installation' }
      responses:
        '201':
          description: |
            สำเร็จ — คืน token ชุดเต็ม
            หรือคืน challenge ถ้าบัญชีนั้นต้องยืนยัน 2FA
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              examples:
                loggedIn:
                  summary: เข้าสู่ระบบสำเร็จ
                  value:
                    success: true
                    data:
                      requiresTwoFactor: false
                      accessToken: eyJhbGciOi...
                      accessTokenExpiresAt: '2026-08-03T04:30:00.000Z'
                      refreshToken: eyJhbGciOi...
                      refreshTokenExpiresAt: '2026-09-02T04:00:00.000Z'
                      tokenType: Bearer
                      session:
                        id: 9f1c...
                        installationId: 6f0d...
                      user:
                        id: usr_1
                        email: employee@example.com
                        displayName: สมชาย ใจดี
                    requestId: req_1
                twoFactorRequired:
                  summary: ต้องยืนยัน 2FA ต่อ
                  value:
                    success: true
                    data:
                      requiresTwoFactor: true
                      twoFactorToken: eyJhbGciOi...
                      expiresAt: '2026-08-03T04:05:00.000Z'
                    requestId: req_2
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/2fa/verify:
    post:
      tags: [Auth]
      summary: ยืนยันรหัส 2FA
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [twoFactorToken, code, installation]
              properties:
                twoFactorToken: { type: string }
                code: { type: string, minLength: 6, maxLength: 6 }
                installation: { $ref: '#/components/schemas/Installation' }
      responses:
        '201':
          description: ยืนยันสำเร็จ คืน token ชุดเต็ม
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessEnvelope' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: ต่ออายุ session
      description: |
        Refresh token หมุนทุกครั้ง (rotation) — ตัวเดิมใช้ต่อไม่ได้ทันที
        `installationId` ต้องตรงกับเครื่องที่สร้าง session ไว้ ไม่อย่างนั้นถูกปฏิเสธ
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken, installationId]
              properties:
                refreshToken: { type: string }
                installationId: { type: string }
      responses:
        '201':
          description: ต่ออายุสำเร็จ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessEnvelope' }
        '401':
          description: token หมดอายุ ถูกเพิกถอน หรือมาจากเครื่องอื่น
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: ออกจากระบบ
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
                installationId: { type: string }
      responses:
        '201':
          description: ออกจากระบบแล้ว (idempotent — เรียกซ้ำได้)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessEnvelope' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: ยังไม่เข้าสู่ระบบ หรือ token ใช้ไม่ได้
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    TooManyRequests:
      description: เรียกบ่อยเกินกำหนด
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, const: true }
        data: {}
        meta: {}
        summary: {}
        requestId: { type: string }
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
        requestId: { type: string }
    Installation:
      type: object
      required: [installationId, platform]
      properties:
        installationId: { type: string, minLength: 8, maxLength: 128 }
        platform: { type: string, enum: [android, ios] }
        deviceName: { type: string }
        deviceModel: { type: string }
        osVersion: { type: string }
        appVersion: { type: string }
        appBuild: { type: integer, minimum: 0 }
        locale: { type: string }
        timezone: { type: string }
...